Added base implementation for switching weapons
else if (isItemAGun(item.itemData))
            {
                if (gunBase != null)
                {
                    gunBase.NewData = item.itemData as GunData;
                    gunBase.EquipGun(gunBase.NewData);
                }
                else
                {
                    Debug.LogError("No GunBase found on the player!");
                }

                RemoveFromIndex(index);
            }
            else
            {
                Debug.Log("Item is not consumable or null");
            }
        }
    }

}
ChatGPT
I apologize for any confusion. If GetComponentInChildren is not finding the GunBase component, there might be an issue with how the GameObject hierarchy is set up. Please make sure that:

The GunBase component is attached to a GameObject that is a child or descendant of the player GameObject.
There are no typos in the name of the GunBase component.
If you've confirmed that the above conditions are met, and you're still encountering issues, you can try the following debugging steps:

Add a Debug.Log statement to check the name of the GameObject that GetComponentInChildren is searching for. This will help you verify that the correct GameObject is being searched for.
csharp
Copy code
GunBase gunBase = GetComponentInChildren<GunBase>();
Debug.Log("Searched for GunBase on: " + gameObject.name);
Verify that the GunBase GameObject is indeed a child or descendant of the player GameObject. You can do this by checking the hierarchy in the Unity Editor.

Double-check that there are no script execution order issues or race conditions causing the GunBase component to be unavailable at the time of the search.

If you're still facing issues after these checks, please provide more information about your GameObject hierarchy and how the GunBase component is set up in your scene so that I can assist you further.

Start with this